# docs/swagger.yaml
basePath: /
definitions:
  http.AddContactRequest:
    properties:
      city:
        type: integer
      country:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      phones:
        items:
          $ref: '#/definitions/model.PhoneType'
        type: array
      postal_code:
        type: string
      street:
        type: string
    required:
    - city
    - country
    - first_name
    - phones
    - street
    type: object
  model.AddressType:
    properties:
      cityId:
        type: integer
      countryId:
        type: integer
      postalCode:
        type: string
      street:
        type: string
    type: object
  model.ContactType:
    properties:
      address:
        $ref: '#/definitions/model.AddressType'
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      phones:
        items:
          $ref: '#/definitions/model.PhoneType'
        type: array
    type: object
  model.PhoneType:
    properties:
      number:
        type: string
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A simple RESTful API for managing a phone book.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Phone Book API
  version: "1.0"
paths:
  /contacts:
    get:
      consumes:
      - application/json
      description: Get paginated contacts from the phone book
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ContactType'
            type: array
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to fetch contacts
          schema:
            type: string
      summary: Get paginated contacts
      tags:
      - contacts
    post:
      consumes:
      - application/json
      description: Add a new contact to the phone book
      parameters:
      - description: Contact to add
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/http.AddContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.AddContactRequest'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Failed to save contact
          schema:
            type: string
      summary: Add a new contact
      tags:
      - contacts
swagger: "2.0"